+++ P2P-Java-desktop/frontend-desktop/src/main/java/tin/p2p/controllerGUI/File.java
@@ -6,13 +6,13 @@ public class File {
     private String name;
     private String ip;
     private String hash;
-    private Integer size;
+    private Long size;
 
     public File(ArrayList<String> fileParams, String filesOwner) {
         // todo obsługa błędu gdy nie wszystkie 4 pola są dostępne
         this.name = fileParams.get(0);
         this.hash = fileParams.get(1);
-        this.size = Integer.valueOf(fileParams.get(2));
+        this.size = Long.valueOf(fileParams.get(2));
         this.ip = filesOwner;
     }
 
@@ -28,7 +28,7 @@ public class File {
         return hash;
     }
 
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
